Allow arbitrary text prefixing
This eliminates the dependency on home-assistant/core#104523, as that PR
does not align with the interpretation of the HASS developers on how STT
should work.

Instead, we just do the work manually in this provider. This requires
some more duplication (of STT providers now, in addition to Assist
pipelines), but is otherwise not particularly intrusive.

One change from this is the addition of the prefix text to the unique ID
of the entity, to ensure that they are properly differentiated.

Also add the potential caveats of this approach to the README.